requirements.txt
streamlit==1.34.0
langchain==0.2.0
langchain-core==0.1.56
langchain-community==0.2.0
langchain-groq==0.1.2
chromadb==0.4.24
sentence-transformers==2.6.1
pypdf==4.1.0
pdfminer.six==20221105
python-dotenv==1.0.1
faiss-cpu
tqdm
nltk
scikit-learn
numpy